Recode

+++ src/Zaabee.StackExchangeRedis/Zaabee.StackExchangeRedis.Abstractions/ISortedSetAsync.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -6,27 +5,27 @@ namespace Zaabee.StackExchangeRedis.Abstractions
 {
     public interface ISortedSetAsync
     {
-        Task<bool> SortedSetAddAsync<T>(string key, T member, long score);
+        Task<bool> SortedSetAddAsync<T>(string key, T member, double score);
 
         Task<long> SortedSetAddAsync<T>(string key, IDictionary<T,double> values);
 
-        Task<double> SortedSetDecrementAsync<T>(string key, T member, long value);
+        Task<double> SortedSetDecrementAsync<T>(string key, T member, double value);
 
-        Task<double> SortedSetIncrementAsync<T>(string key, T member, long value);
+        Task<double> SortedSetIncrementAsync<T>(string key, T member, double value);
 
         Task<long> SortedSetLengthAsync<T>(string key);
 
         Task<long> SortedSetLengthByValueAsync<T>(string key, T min, T max);
 
-        Task<IList<T>> SortedSetRangeByScoreAscendingAsync<T>(string key, long start = 0, long stop = -1);
+        Task<IList<T>> SortedSetRangeByScoreAscendingAsync<T>(string key, double start = 0, double stop = -1);
 
-        Task<IList<T>> SortedSetRangeByScoreDescendingAsync<T>(string key, long start = 0, long stop = -1);
+        Task<IList<T>> SortedSetRangeByScoreDescendingAsync<T>(string key, double start = 0, double stop = -1);
 
         Task<IDictionary<T,double>> SortedSetRangeByScoreWithScoresAscendingAsync<T>(string key, long start = 0,
             long stop = -1);
 
-        Task<IDictionary<T,double>> SortedSetRangeByScoreWithScoresDescendingAsync<T>(string key, long start = 0,
-            long stop = -1);
+        Task<IDictionary<T,double>> SortedSetRangeByScoreWithScoresDescendingAsync<T>(string key, double start = 0,
+            double stop = -1);
 
         Task<IList<T>> SortedSetRangeByValueAsync<T>(string key, T min, T max, long skip, long take = -1);
 
@@ -42,7 +41,7 @@ namespace Zaabee.StackExchangeRedis.Abstractions
 
         Task<long> SortedSetRemoveRangeAsync<T>(string key, IEnumerable<T> members);
 
-        Task<long> SortedSetRemoveRangeByScoreAsync<T>(string key, long start, long stop);
+        Task<long> SortedSetRemoveRangeByScoreAsync<T>(string key, double start, double stop);
 
         Task<long> SortedSetRemoveRangeByValueAsync<T>(string key, T min, T max);
